+++ tests/num_digits_tests.cpp
@@ -8,31 +8,31 @@ using jewel::NumDigits;
 
 TEST(num_digits_with_default_base_int)
 {
-	CHECK_EQUAL(NumDigits::num_digits(0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(0, 10), 1);
-	CHECK_EQUAL(NumDigits::num_digits(1), 1);
-	CHECK_EQUAL(NumDigits::num_digits(-1), 1);
-	CHECK_EQUAL(NumDigits::num_digits(2), 1);
-	CHECK_EQUAL(NumDigits::num_digits(6), 1);
-	CHECK_EQUAL(NumDigits::num_digits(-6), 1);
-	CHECK_EQUAL(NumDigits::num_digits(6, 10), 1);
-	CHECK_EQUAL(NumDigits::num_digits(-6, 10), 1);
-	CHECK_EQUAL(NumDigits::num_digits(9), 1);
-	CHECK_EQUAL(NumDigits::num_digits(-9), 1);
-	CHECK_EQUAL(NumDigits::num_digits(10), 2);
-	CHECK_EQUAL(NumDigits::num_digits(-10), 2);
-	CHECK_EQUAL(NumDigits::num_digits(34), 2);
-	CHECK_EQUAL(NumDigits::num_digits(-34), 2);
-	CHECK_EQUAL(NumDigits::num_digits(100), 3);
-	CHECK_EQUAL(NumDigits::num_digits(-100), 3);
-	CHECK_EQUAL(NumDigits::num_digits(999), 3);
-	CHECK_EQUAL(NumDigits::num_digits(-999), 3);
-	CHECK_EQUAL(NumDigits::num_digits(1000), 4);
-	CHECK_EQUAL(NumDigits::num_digits(-1000), 4);
-	CHECK_EQUAL(NumDigits::num_digits(10798778), 8);
-	CHECK_EQUAL(NumDigits::num_digits(-10798778), 8);
-	CHECK_EQUAL(NumDigits::num_digits(999999999), 9);
-	CHECK_EQUAL(NumDigits::num_digits(-999999999), 9);
+	CHECK_EQUAL(NumDigits::num_digits(0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(0, 10), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(1), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(-1), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(2), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(6), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(-6), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(6, 10), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(-6, 10), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(9), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(-9), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(10), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(-10), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(34), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(-34), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(100), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(-100), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(999), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(-999), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(1000), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(-1000), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(10798778), static_cast<size_t>(8));
+	CHECK_EQUAL(NumDigits::num_digits(-10798778), static_cast<size_t>(8));
+	CHECK_EQUAL(NumDigits::num_digits(999999999), static_cast<size_t>(9));
+	CHECK_EQUAL(NumDigits::num_digits(-999999999), static_cast<size_t>(9));
 	CHECK_EQUAL
 	(	NumDigits::num_digits(INT_MAX),
 		NumDigits::num_digits(INT_MIN)
@@ -41,15 +41,15 @@ TEST(num_digits_with_default_base_int)
 
 TEST(num_digits_with_base_of_2_int)
 {
-	CHECK_EQUAL(NumDigits::num_digits(0, 2), 1);
-	CHECK_EQUAL(NumDigits::num_digits(1, 2), 1);
-	CHECK_EQUAL(NumDigits::num_digits(-1, 2), 1);
-	CHECK_EQUAL(NumDigits::num_digits(2, 2), 2);
-	CHECK_EQUAL(NumDigits::num_digits(-2, 2), 2);
-	CHECK_EQUAL(NumDigits::num_digits(3, 2), 2);
-	CHECK_EQUAL(NumDigits::num_digits(4, 2), 3);
-	CHECK_EQUAL(NumDigits::num_digits(7, 2), 3);
-	CHECK_EQUAL(NumDigits::num_digits(-8, 2), 4);
+	CHECK_EQUAL(NumDigits::num_digits(0, 2), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(1, 2), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(-1, 2), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(2, 2), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(-2, 2), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(3, 2), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(4, 2), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(7, 2), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(-8, 2), static_cast<size_t>(4));
 }
 
 TEST(num_digits_with_base_of_10_short)
@@ -60,12 +60,12 @@ TEST(num_digits_with_base_of_10_short)
 	short i1243= 1243;
 	short i_0 = -0;
 	short i_12342 = -2343;
-	CHECK_EQUAL(NumDigits::num_digits(i0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(i10), 2);
-	CHECK_EQUAL(NumDigits::num_digits(i123), 3);
-	CHECK_EQUAL(NumDigits::num_digits(i1243), 4);
-	CHECK_EQUAL(NumDigits::num_digits(i_0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(i_12342), 4);
+	CHECK_EQUAL(NumDigits::num_digits(i0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(i10), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(i123), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(i1243), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(i_0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(i_12342), static_cast<size_t>(4));
 }
 
 TEST(num_digits_with_base_of_10_long)
@@ -77,13 +77,13 @@ TEST(num_digits_with_base_of_10_long)
 	long i_0 = -0;
 	long i_12342 = -2343;
 	long i_976926943 = -976926943;
-	CHECK_EQUAL(NumDigits::num_digits(i0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(i10), 2);
-	CHECK_EQUAL(NumDigits::num_digits(i123), 3);
-	CHECK_EQUAL(NumDigits::num_digits(i1243), 4);
-	CHECK_EQUAL(NumDigits::num_digits(i_0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(i_12342), 4);
-	CHECK_EQUAL(NumDigits::num_digits(i_976926943), 9); 
+	CHECK_EQUAL(NumDigits::num_digits(i0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(i10), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(i123), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(i1243), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(i_0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(i_12342), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(i_976926943), static_cast<size_t>(9)); 
 }
 
 TEST(num_digits_with_base_of_10_unsigned_int)
@@ -93,11 +93,11 @@ TEST(num_digits_with_base_of_10_unsigned_int)
 	unsigned int i123 = 123;
 	unsigned int i1243 = 1243;
 	unsigned int i887423 = 887423;
-	CHECK_EQUAL(NumDigits::num_digits(i0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(i10), 2);
-	CHECK_EQUAL(NumDigits::num_digits(i123), 3);
-	CHECK_EQUAL(NumDigits::num_digits(i1243), 4);
-	CHECK_EQUAL(NumDigits::num_digits(i887423), 6);
+	CHECK_EQUAL(NumDigits::num_digits(i0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(i10), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(i123), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(i1243), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(i887423), static_cast<size_t>(6));
 }
 
 TEST(num_digits_with_base_of_10_unsigned_short)
@@ -107,11 +107,11 @@ TEST(num_digits_with_base_of_10_unsigned_short)
 	unsigned short i123 = 123;
 	unsigned short i1243 = 1243;
 	unsigned short i9762 = 9762;
-	CHECK_EQUAL(NumDigits::num_digits(i0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(i10), 2);
-	CHECK_EQUAL(NumDigits::num_digits(i123), 3);
-	CHECK_EQUAL(NumDigits::num_digits(i1243), 4);
-	CHECK_EQUAL(NumDigits::num_digits(i9762), 4);
+	CHECK_EQUAL(NumDigits::num_digits(i0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(i10), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(i123), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(i1243), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(i9762), static_cast<size_t>(4));
 }
 
 TEST(num_digits_with_base_of_10_unsigned_long)
@@ -121,11 +121,11 @@ TEST(num_digits_with_base_of_10_unsigned_long)
 	unsigned long i123 = 123;
 	unsigned long i1243 = 1243;
 	unsigned long i887423 = 887423;
-	CHECK_EQUAL(NumDigits::num_digits(i0), 1);
-	CHECK_EQUAL(NumDigits::num_digits(i10), 2);
-	CHECK_EQUAL(NumDigits::num_digits(i123), 3);
-	CHECK_EQUAL(NumDigits::num_digits(i1243), 4);
-	CHECK_EQUAL(NumDigits::num_digits(i887423), 6);
+	CHECK_EQUAL(NumDigits::num_digits(i0), static_cast<size_t>(1));
+	CHECK_EQUAL(NumDigits::num_digits(i10), static_cast<size_t>(2));
+	CHECK_EQUAL(NumDigits::num_digits(i123), static_cast<size_t>(3));
+	CHECK_EQUAL(NumDigits::num_digits(i1243), static_cast<size_t>(4));
+	CHECK_EQUAL(NumDigits::num_digits(i887423), static_cast<size_t>(6));
 }
